[refactor] Move code around; no-op change.
This sets us up for implementing better enum support.
